Added non reactive, unsteady catalyticPimpleFOAM solver (version 0.9)
The catalyticPimpleFOAM solver was added without the homogeneous and heterogeneous reactions (to be added as soon as possible). The catalyticPimpleFOAM solver now uses the current OpenSMOKE++ library, with the most recent features. The solver was derived from the reactingFOAM solver available in the OpenFOAM 2.3.x version. It is based on the PIMPLE loop, which means that the fvSolution file needs to be adapted accordingly (the previous versions of catalyticFOAM were based on the PISO methodology). No post-processing capabilities were included in the new solver (elelemntal species balances, conversions, monitoring of selected species) in order to increase the efficiency as much as possible.

The solution of the chemical step (through the solution of stiff ODE solvers) will be added in the next future.

A simple 2D case (channel fed with H2 and O2) was added, together the chemical interpreter for the kinetic mechanisms with surface reactions.